+++ build.sh
@@ -119,22 +119,46 @@ build_altbn128() {
   ############################
 EOF
 
+  echo "building altbn128 for ${OSARCH}"
   cd "$SCRIPTDIR/altbn128/sputnikvm_altbn128"
 
   # delete old build dir, if exists
   rm -rf "$SCRIPTDIR/altbn128/build" || true
-  mkdir -p "$SCRIPTDIR/altbn128/build/lib"
+  mkdir -p "$SCRIPTDIR/altbn128/build/${OSARCH}/lib"
 
   cargo clean
-
   if [[ "$OSTYPE" == "darwin"* ]];  then
     lipo_lib "libeth_altbn128" ""
   else
     cargo build --lib --release
   fi
-
   mkdir -p "$SCRIPTDIR/altbn128/build/${OSARCH}/lib"
   cp target/release/libeth_altbn128.* "$SCRIPTDIR/altbn128/build/${OSARCH}/lib"
+}
+
+build_arithmetic() {
+  cat <<EOF
+  ##############################
+  ###### build arithmetic ######
+  ##############################
+EOF
+
+  cd "$SCRIPTDIR/arithmetic/arithmetic"
+
+  # delete old build dir, if exists
+  rm -rf "$SCRIPTDIR/arithmetic/build" || true
+  mkdir -p "$SCRIPTDIR/arithmetic/build/lib"
+
+  cargo clean
+
+  if [[ "$OSTYPE" == "darwin"* ]];  then
+    lipo_lib "libeth_arithmetic" ""
+  else
+    cargo build --lib --release
+  fi
+
+  mkdir -p "$SCRIPTDIR/arithmetic/build/${OSARCH}/lib"
+  cp target/release/libeth_arithmetic.* "$SCRIPTDIR/arithmetic/build/${OSARCH}/lib"
 }
 
 build_ipa_multipoint() {
@@ -290,6 +314,7 @@ EOF
 build_blake2bf
 build_secp256k1
 build_altbn128
+build_arithmetic
 build_bls12_381
 build_ipa_multipoint
 build_secp256r1